refactor: split Core.swift into one file per type
Core.swift was a grab-bag of ~10 unrelated public types. Split it so each
file is named after the type it contains, and every same-file extension
(protocol default impls, the LocalizedError conformance, firstPresentKey)
travels with its type rather than being hoisted into a +Feature file —
those extensions are part of their type's definition, not cross-cutting
features on a foreign type.

SwiftSync, SyncModelable (+ defaults), SyncRelationshipSchemaDescriptor,
SyncUpdatableModel (+ defaults), ParentScopedModel, SyncRelationshipLookupCache
(+ lookup state), SyncRelationshipOperations, KeyStyle, ExportState (+ box),
SyncPayloadConvertible, SyncPayload (+ candidate-keys cache, firstPresentKey),
SyncError (+ LocalizedError). Stale MARK headers dropped.

Pure code move: SwiftSync 172 (+48), DemoCore unaffected — green.
