+++ SwiftSync/Sources/SwiftSync/ParentScopedModel.swift
@@ -1,0 +1,6 @@
+import SwiftData
+
+public protocol ParentScopedModel: SyncUpdatableModel {
+    associatedtype SyncParent: PersistentModel
+    static var parentRelationship: ReferenceWritableKeyPath<Self, SyncParent?> { get }
+}
