+++ SwiftSync/Sources/SwiftSync/SwiftSync.swift
@@ -1,0 +1,1 @@
+public enum SwiftSync {}
