+++ SwiftSync/Sources/SwiftSync/SyncPayloadConvertible.swift
@@ -1,0 +1,3 @@
+public protocol SyncPayloadConvertible: Sendable {
+    func toSyncPayloadDictionary() -> [String: Any]
+}
